# sinistros_back/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Cadastro de Sinistros",
    "description": "Documentação da API de cadastro de sinistros",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/corretora/": {
      "get": {
        "tags": [
          "Corretoras"
        ],
        "summary": "Lista todas as corretoras",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Corretoras"
        ],
        "summary": "Cadastra uma corretora",
        "description": "Realiza o cadastro de uma nova corretora.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "Corretora ABC"
                },
                "corretorResponsavel": {
                  "type": "string",
                  "example": "Maria"
                },
                "email": {
                  "type": "string",
                  "example": "contato@corretora.com"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Corretora cadastrada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          }
        }
      }
    },
    "/corretora/{id}": {
      "delete": {
        "tags": [
          "Corretoras"
        ],
        "summary": "Remove uma corretora",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID da corretora a ser removida"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "Corretoras"
        ],
        "summary": "Atualiza uma corretora",
        "description": "Atualiza os dados de uma corretora existente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID da corretora a ser atualizada"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "Corretora ABC"
                },
                "corretorResponsavel": {
                  "type": "string",
                  "example": "Maria"
                },
                "email": {
                  "type": "string",
                  "example": "contato@corretora.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Corretora atualizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          }
        }
      }
    },
    "/corretora-cia/": {
      "get": {
        "tags": [
          "Corretora x Cia"
        ],
        "summary": "Lista os vínculos entre corretoras e cias de seguro",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Corretora x Cia"
        ],
        "summary": "Cria um vínculo entre corretora e cia de seguro",
        "description": "Vincula uma corretora a uma cia de seguro.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "corretoraId": {
                  "type": "number",
                  "example": 1
                },
                "ciaId": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Vínculo criado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          }
        }
      }
    },
    "/corretora-cia/{corretoraId}/{ciaId}": {
      "delete": {
        "tags": [
          "Corretora x Cia"
        ],
        "summary": "Remove um vínculo entre corretora e cia de seguro",
        "description": "",
        "parameters": [
          {
            "name": "corretoraId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID da corretora"
          },
          {
            "name": "ciaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID da cia de seguro"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/sinistro/": {
      "get": {
        "tags": [
          "Sinistros"
        ],
        "summary": "Lista todos os sinistros",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Sinistros"
        ],
        "summary": "Cadastra um sinistro",
        "description": "Realiza o cadastro de um novo sinistro. Se o tipo de atendimento for SEGURO, é obrigatório informar a cia de seguro e a corretora.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "veiculoId": {
                  "type": "number",
                  "example": 1
                },
                "tipoAtendimento": {
                  "type": "string",
                  "example": "SEGURO"
                },
                "ciaSeguroId": {
                  "type": "number",
                  "example": 1
                },
                "corretoraId": {
                  "type": "number",
                  "example": 1
                },
                "numApolice": {
                  "type": "string",
                  "example": "123456"
                },
                "kmAtendimento": {
                  "type": "number",
                  "example": 15000
                },
                "statusAtual": {
                  "type": "string",
                  "example": "Em andamento"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Sinistro cadastrado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          }
        }
      }
    },
    "/sinistro/{id}": {
      "delete": {
        "tags": [
          "Sinistros"
        ],
        "summary": "Remove um sinistro",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do sinistro a ser removido"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "Sinistros"
        ],
        "summary": "Atualiza um sinistro",
        "description": "Atualiza os dados de um sinistro existente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do sinistro a ser atualizado"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "veiculoId": {
                  "type": "number",
                  "example": 1
                },
                "tipoAtendimento": {
                  "type": "string",
                  "example": "SEGURO"
                },
                "ciaSeguroId": {
                  "type": "number",
                  "example": 1
                },
                "corretoraId": {
                  "type": "number",
                  "example": 1
                },
                "numApolice": {
                  "type": "string",
                  "example": "123456"
                },
                "kmAtendimento": {
                  "type": "number",
                  "example": 15000
                },
                "statusAtual": {
                  "type": "string",
                  "example": "Em andamento"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sinistro atualizado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          }
        }
      }
    },
    "/historico-sinistro/": {
      "get": {
        "tags": [
          "Histórico de Sinistros"
        ],
        "summary": "Lista todos os históricos de sinistro",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Histórico de Sinistros"
        ],
        "summary": "Cadastra um histórico de sinistro",
        "description": "Registra uma nova entrada no histórico de um sinistro.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "sinistroId": {
                  "type": "number",
                  "example": 1
                },
                "status": {
                  "type": "string",
                  "example": "Em andamento"
                },
                "observacao": {
                  "type": "string",
                  "example": "Veículo em análise"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Histórico cadastrado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          }
        }
      }
    },
    "/historico-sinistro/{id}": {
      "put": {
        "tags": [
          "Histórico de Sinistros"
        ],
        "summary": "Atualiza um histórico de sinistro",
        "description": "Atualiza os dados de um histórico de sinistro existente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do histórico a ser atualizado"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "sinistroId": {
                  "type": "number",
                  "example": 1
                },
                "status": {
                  "type": "string",
                  "example": "Em andamento"
                },
                "observacao": {
                  "type": "string",
                  "example": "Veículo em análise"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico atualizado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          }
        }
      },
      "delete": {
        "tags": [
          "Histórico de Sinistros"
        ],
        "summary": "Remove um histórico de sinistro",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do histórico a ser removido"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  }
}